Add an undocumented -n option to skip non-free wifi and other firmware installation by fw_update

+++ mkchroot.sh
@@ -9,15 +9,17 @@ mount -t tmpfs -o noatime,nodev,nosuid,noexec tmpfs ${BUILDPATH}/fs/tmp
 
 # Run fw_update, set up a temporary resolv.conf if necessary
 
-echo 'Running fw_update...'
-${ESCECHO} "nameserver 8.8.8.8\nlookup file bind" > ${BUILDPATH}/fs/etc/resolv.conf
-if ! chroot ${BUILDPATH}/fs fw_update -a >> ${BUILDPATH}/05.mkchroot.fw_update 2>&1; then
-  # Handle both arch and snapshot binary incompatibilities
-  fw_update -a >> ${BUILDPATH}/05.mkchroot.fw_update 2>&1
-  cp -fRPp /etc/firmware/* ${BUILDPATH}/fs/etc/firmware
-  cp -fRPp /var/db/pkg/*-firmware-* ${BUILDPATH}/fs/var/db/pkg
+if [ -z "${skipfw+1}" ]; then
+  echo 'Running fw_update...'
+  ${ESCECHO} "nameserver 8.8.8.8\nlookup file bind" > ${BUILDPATH}/fs/etc/resolv.conf
+  if ! chroot ${BUILDPATH}/fs fw_update -a >> ${BUILDPATH}/05.mkchroot.fw_update 2>&1; then
+    # Handle both arch and snapshot binary incompatibilities
+    fw_update -a >> ${BUILDPATH}/05.mkchroot.fw_update 2>&1
+    cp -fRPp /etc/firmware/* ${BUILDPATH}/fs/etc/firmware
+    cp -fRPp /var/db/pkg/*-firmware-* ${BUILDPATH}/fs/var/db/pkg
+  fi
+  rm ${BUILDPATH}/fs/etc/resolv.conf
 fi
-rm ${BUILDPATH}/fs/etc/resolv.conf
 
 # Install packages, if directed
